# backend/app/routes/gallery.py
"""
Gallery routes with API-side pagination, cursor-based pagination,
sorting, field selection, and ETag support.
"""

from typing import Annotated, Any
from uuid import UUID

from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Path, Query, Request, Response
from pydantic import ValidationError

from app.dependencies import get_current_token, get_gallery_service, get_storage_service
from app.limiter import get_api_limit, limiter
from app.logger import logger, sanitize_log_value
from app.middleware.auth_middleware import (
    get_current_user_from_credentials,
    get_current_user_optional,
)
from app.schemas.gallery import (
    GALLERY_ALLOWED_FIELDS,
    GalleryResponse,
    PaginatedGalleryResponse,
    PaginationMeta,
    PopularTagsResponse,
    SearchResponse,
    SortField,
    SortOrder,
    TagInfo,
)
from app.schemas.location import CatLocation
from app.schemas.user import User
from app.services.gallery_service import GalleryService
from app.services.storage_service import StorageService
from app.utils.location_utils import protect_photo_location, protect_photo_locations

router = APIRouter(prefix="/gallery", tags=["Gallery"])

PhotoIdPath = Annotated[UUID, Path(title="The ID of the photo", description="Must be a valid UUID")]


def _calculate_offset(offset: int, page: int | None, limit: int) -> int:
    """Calculate actual pagination offset based on offset or 1-based page number."""
    if page is not None:
        return (page - 1) * limit
    return offset


async def schedule_photo_deletion(
    photo_id_str: str,
    user_id: str,
    gallery_service: GalleryService,
    storage_service: StorageService,
    background_tasks: BackgroundTasks,
) -> None:
    """Verify ownership and schedule photo deletion in background."""
    photo = await gallery_service.verify_photo_ownership(photo_id_str, user_id)
    if not photo:
        raise HTTPException(status_code=404, detail="Photo not found or access denied")

    background_tasks.add_task(
        gallery_service.process_photo_deletion,
        photo_id=photo_id_str,
        image_url=photo.get("image_url") or "",
        user_id=user_id,
        storage_service=storage_service,
    )


def _filter_fields(data: dict[str, Any], fields: set[str] | None) -> dict[str, Any]:
    """Filter a dict to only include specified fields."""
    if not fields:
        return data
    return {k: v for k, v in data.items() if k in fields}


def _build_gallery_locations(photos: list[dict[str, Any]]) -> list[CatLocation]:
    """Keep malformed cached or legacy rows from breaking the whole gallery response."""
    locations: list[CatLocation] = []
    skipped = 0
    for photo in photos:
        try:
            locations.append(CatLocation(**photo))
        except ValidationError:
            skipped += 1

    if skipped:
        logger.warning("Skipped %d gallery photos with incomplete location data", skipped)
    return locations


def _selected_gallery_fields(fields: str | None) -> set[str] | None:
    if not fields:
        return None
    requested = {field.strip() for field in fields.split(",") if field.strip()}
    selected = requested & GALLERY_ALLOWED_FIELDS
    selected.update({"id", "image_url", "latitude", "longitude"})
    return selected


def _empty_gallery_response(result: dict[str, Any], offset: int, limit: int) -> PaginatedGalleryResponse:
    total = int(result.get("total") or 0)
    current_page = (offset // limit) + 1 if limit > 0 else 1
    total_pages = (total + limit - 1) // limit if total > 0 else 0
    return PaginatedGalleryResponse(
        images=[],
        pagination=PaginationMeta(
            total=total,
            limit=limit,
            offset=offset,
            has_more=bool(result.get("has_more", False)),
            page=current_page,
            total_pages=total_pages,
        ),
    )


def _apply_sort(
    photos: list[dict[str, Any]],
    sort: SortField | None,
    order: SortOrder,
) -> list[dict[str, Any]]:
    """Sort photos list by the given field and order."""
    if not sort:
        return photos

    reverse = order == SortOrder.DESC
    sort_key = sort.value

    def _get_sort_val(item: dict[str, Any]) -> Any:
        val = item.get(sort_key)
        if val is None:
            # Sort NULLs consistently at the end for all types
            if sort_key == SortField.UPLOADED_AT:
                return "0000-00-00"
            return -1
        return val

    return sorted(photos, key=_get_sort_val, reverse=reverse)


# ---- Offset-based pagination (default) ----


@router.get(
    "",
    responses={
        200: {"description": "Successful Response"},
        500: {"description": "Internal Server Error"},
    },
)
@limiter.limit(get_api_limit)
async def get_gallery(
    request: Request,
    response: Response,
    gallery_service: Annotated[GalleryService, Depends(get_gallery_service)],
    current_user: Annotated[User | None, Depends(get_current_user_optional)],
    token: Annotated[str | None, Depends(get_current_token)],
    limit: Annotated[int, Query(ge=1, le=100, description="Number of items per page")] = 20,
    offset: Annotated[int, Query(ge=0, description="Number of items to skip")] = 0,
    page: Annotated[int | None, Query(ge=1, description="Page number (alternative to offset)")] = None,
    sort: Annotated[SortField | None, Query(description="Sort field: uploaded_at, likes_count, comments_count")] = None,
    order: Annotated[SortOrder, Query(description="Sort order: asc or desc")] = SortOrder.DESC,
    fields: Annotated[str | None, Query(description="Comma-separated list of fields to include")] = None,
) -> PaginatedGalleryResponse:
    """
    Get cat images with pagination, sorting, and field selection.

    Examples:
    - `/gallery?limit=20&offset=0` - First 20 images
    - `/gallery?limit=20&page=2` - Second page of 20 images
    - `/gallery?sort=likes_count&order=desc` - Sort by most liked
    - `/gallery?fields=id,image_url,location_name` - Only specific fields
    """
    # Dynamic caching strategy
    if current_user:
        response.headers["Cache-Control"] = "no-cache, no-store, must-revalidate"
    else:
        # Improved caching for public gallery
        response.headers["Cache-Control"] = "public, max-age=300, stale-while-revalidate=60"

    selected_fields = _selected_gallery_fields(fields)

    try:
        actual_offset = _calculate_offset(offset, page, limit)

        result = await gallery_service.get_all_photos(
            limit=limit,
            offset=actual_offset,
            include_total=True,
            user_id=current_user.id if current_user else None,
            jwt_token=token if current_user else None,
            sort_field=sort.value if sort else None,
            sort_desc=(order == SortOrder.DESC),
        )

        if not result["data"]:
            return _empty_gallery_response(result, actual_offset, limit)

        # Sorting is now handled at the DB level via sort_field/sort_desc
        protected_data = protect_photo_locations(result["data"])

        # Apply field selection
        if selected_fields:
            protected_data = [_filter_fields(d, selected_fields) for d in protected_data]

        total = result["total"]
        total_pages = (total + limit - 1) // limit if total > 0 else 0
        current_page = (actual_offset // limit) + 1 if limit > 0 else 1

        cat_locations = _build_gallery_locations(protected_data)

        return PaginatedGalleryResponse(
            images=cat_locations,
            pagination=PaginationMeta(
                total=total,
                limit=limit,
                offset=actual_offset,
                has_more=result["has_more"],
                page=current_page,
                total_pages=total_pages,
            ),
        )

    except Exception as e:
        logger.error("Gallery fetch error: %s", str(e), exc_info=True)
        raise HTTPException(status_code=500, detail="Failed to fetch gallery images")


# ---- Locations endpoint ----


@router.get("/locations", responses={500: {"description": "Internal Server Error"}})
async def get_locations(
    response: Response,
    gallery_service: Annotated[GalleryService, Depends(get_gallery_service)],
    limit: Annotated[int, Query(ge=1, le=500, description="Maximum number of legacy marker results")] = 500,
) -> list[CatLocation]:
    """Get a bounded legacy marker list from Supabase."""
    response.headers["Cache-Control"] = "public, max-age=300"

    try:
        photos = await gallery_service.get_map_locations(limit=limit)
        if not photos:
            return []
        photos = protect_photo_locations(photos)
        return [CatLocation(**photo) for photo in photos]
    except Exception as e:
        logger.error("Locations fetch error: %s", str(e), exc_info=True)
        raise HTTPException(status_code=500, detail="Failed to fetch cat locations")


# ---- Viewport endpoint ----


@router.get("/viewport", responses={500: {"description": "Internal Server Error"}})
async def get_locations_in_viewport(
    response: Response,
    gallery_service: Annotated[GalleryService, Depends(get_gallery_service)],
    current_user: Annotated[User | None, Depends(get_current_user_optional)],
    north: Annotated[float, Query(description="North latitude bound")],
    south: Annotated[float, Query(description="South latitude bound")],
    east: Annotated[float, Query(description="East longitude bound")],
    west: Annotated[float, Query(description="West longitude bound")],
    limit: Annotated[int, Query(ge=1, le=500, description="Maximum number of results")] = 100,
) -> GalleryResponse:
    """Get cat locations within a geographic viewport (bounding box)."""
    # PERF: Enable caching for viewport data so ETag middleware can work
    if current_user:
        response.headers["Cache-Control"] = "private, max-age=60, stale-while-revalidate=30"
    else:
        response.headers["Cache-Control"] = "public, max-age=60, stale-while-revalidate=30"
    try:
        photos = await gallery_service.get_viewport_photos(
            north=north,
            south=south,
            east=east,
            west=west,
            limit=limit,
        )

        if not photos:
            return GalleryResponse(images=[])

        if current_user:
            photos = await gallery_service.enrich_with_user_data(photos, current_user.id)

        photos = protect_photo_locations(photos)
        cat_locations = [CatLocation(**photo) for photo in photos]
        return GalleryResponse(images=cat_locations)

    except Exception as e:
        logger.error("Viewport fetch error: %s", str(e), exc_info=True)
        raise HTTPException(status_code=500, detail="Failed to fetch locations in viewport")


# ---- Search endpoint ----


@router.get("/search", responses={500: {"description": "Internal Server Error"}})
@limiter.limit(get_api_limit)
async def search_locations(
    request: Request,
    gallery_service: Annotated[GalleryService, Depends(get_gallery_service)],
    current_user: Annotated[User | None, Depends(get_current_user_optional)],
    q: Annotated[str | None, Query(description="Text to search in location name and description")] = None,
    tags: Annotated[str | None, Query(description="Comma-separated list of tags to filter by")] = None,
    limit: Annotated[int, Query(ge=1, le=500, description="Maximum number of results")] = 100,
    offset: Annotated[int, Query(ge=0, description="Number of items to skip")] = 0,
    page: Annotated[int | None, Query(ge=1, description="Page number (alternative to offset)")] = None,
    sort: Annotated[SortField | None, Query(description="Sort field")] = None,
    order: Annotated[SortOrder, Query(description="Sort order")] = SortOrder.DESC,
) -> SearchResponse:
    """Search cat locations with optional text query and/or tag filters."""
    try:
        tag_list = None
        if tags:
            tag_list = [t.strip() for t in tags.split(",") if t.strip()]

        actual_offset = _calculate_offset(offset, page, limit)

        photos = await gallery_service.search_photos(
            query=q,
            tags=tag_list,
            limit=limit,
            offset=actual_offset,
            user_id=current_user.id if current_user else None,
            include_total=True,
        )

        # Apply sorting
        if photos:
            photos = _apply_sort(photos, sort, order)
            photos = protect_photo_locations(photos)

        results = [CatLocation(**photo) for photo in photos] if photos else []

        total = int(getattr(photos, "total", len(results)))
        return SearchResponse(
            results=results,
            total=total,
            query=q,
            tags=tag_list,
            limit=limit,
            offset=actual_offset,
        )

    except Exception as e:
        logger.error("Search error: %s", str(e), exc_info=True)
        raise HTTPException(status_code=500, detail="Failed to search locations")


# ---- Popular tags endpoint ----


@router.get("/popular-tags", responses={500: {"description": "Internal Server Error"}})
@limiter.limit(get_api_limit)
async def get_popular_tags(
    request: Request,
    response: Response,
    gallery_service: Annotated[GalleryService, Depends(get_gallery_service)],
    limit: Annotated[int, Query(ge=1, le=100, description="Number of top tags to return")] = 20,
) -> PopularTagsResponse:
    """Get the most popular tags used across all cat photos."""
    response.headers["Cache-Control"] = "public, max-age=3600"

    try:
        tags_data = await gallery_service.get_popular_tags(limit=limit)
        tags = [TagInfo(**t) for t in tags_data]
        return PopularTagsResponse(tags=tags)
    except Exception as e:
        logger.error("Popular tags error: %s", str(e), exc_info=True)
        raise HTTPException(status_code=500, detail="Failed to fetch popular tags")


# ---- Single photo endpoint ----


@router.get(
    "/{photo_id}",
    responses={404: {"description": "Photo not found"}, 500: {"description": "Internal Server Error"}},
)
@limiter.limit(get_api_limit)
async def get_photo(
    request: Request,
    photo_id: PhotoIdPath,
    gallery_service: Annotated[GalleryService, Depends(get_gallery_service)],
    current_user: Annotated[User | None, Depends(get_current_user_optional)],
) -> CatLocation:
    """Get a specific photo by its ID."""
    photo_id_str = str(photo_id)
    try:
        photo = await gallery_service.get_photo_by_id(photo_id_str)
        if not photo:
            raise HTTPException(status_code=404, detail="Photo not found")

        if current_user and photo:
            enriched = await gallery_service.enrich_with_user_data([photo], current_user.id)
            if enriched:
                photo = enriched[0]

        photo = protect_photo_location(photo)
        return CatLocation(**photo)
    except HTTPException:
        raise
    except Exception as e:
        logger.error("Error fetching photo %s: %s", sanitize_log_value(photo_id_str), e, exc_info=True)
        raise HTTPException(status_code=500, detail="Failed to fetch photo")


# ---- Delete photo endpoint ----


@router.delete(
    "/{photo_id}",
    status_code=202,
    responses={404: {"description": "Photo not found or access denied"}, 500: {"description": "Internal Server Error"}},
)
async def delete_photo(
    photo_id: PhotoIdPath,
    background_tasks: BackgroundTasks,
    current_user: Annotated[User, Depends(get_current_user_from_credentials)],
    gallery_service: Annotated[GalleryService, Depends(get_gallery_service)],
    storage_service: Annotated[StorageService, Depends(get_storage_service)],
) -> dict[str, str]:
    """Delete a photo. Returns 202 Accepted (deletion runs in background)."""
    photo_id_str = str(photo_id)
    try:
        await schedule_photo_deletion(photo_id_str, current_user.id, gallery_service, storage_service, background_tasks)
        return {"message": "Deletion scheduled"}

    except HTTPException:
        raise
    except Exception as e:
        logger.error("Delete request failed: %s", str(e), exc_info=True)
        raise HTTPException(status_code=500, detail="Failed to schedule deletion")