# docs/openapi-baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PurrFect Spots API",
    "description": "\n        PurrFect Spots API helps you share and discover cat locations in a socially-connected way.\n\n        ### 🌐 API Specification\n        All endpoints follow the `/api/v1/` prefix convention for stable identification.\n\n        ### ✨ Key Features\n        * 📍 **Geo-Location Awareness**: High-precision cat spot sharing with privacy protection.\n        * 🤳 **AI Photo Intelligence**: Automated cat detection and visual attribute extraction.\n        * 🔒 **Enterprise-Grade Security**: PKCE-enabled OAuth 2.0 and Supabase-backed authentication.\n        * ⚡ **Performance Optimized**: Built-in rate limiting, GZip compression, and ETag support.\n        * 🔄 **Real-time Ready**: Designed for low-latency interactions and live updates.\n        ",
    "contact": {
      "name": "Purrfect Spots Team",
      "email": "support@purrfectspots.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "3.1.0"
  },
  "paths": {
...
    "/api/v1/gallery": {
      "get": {
        "tags": [
          "Gallery"
        ],
        "summary": "Get Gallery",
        "description": "Get cat images with pagination, sorting, and field selection.\n\nExamples:\n- `/gallery?limit=20&offset=0` - First 20 images\n- `/gallery?limit=20&page=2` - Second page of 20 images\n- `/gallery?sort=likes_count&order=desc` - Sort by most liked\n- `/gallery?fields=id,image_url,location_name` - Only specific fields",
        "operationId": "get_gallery_api_v1_gallery_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page number (alternative to offset)",
              "title": "Page"
            },
            "description": "Page number (alternative to offset)"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SortField"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort field: uploaded_at, likes_count, comments_count",
              "title": "Sort"
            },
            "description": "Sort field: uploaded_at, likes_count, comments_count"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Sort order: asc or desc",
              "default": "desc"
            },
            "description": "Sort order: asc or desc"
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of fields to include",
              "title": "Fields"
            },
            "description": "Comma-separated list of fields to include"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGalleryResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gallery/locations": {
      "get": {
        "tags": [
          "Gallery"
        ],
        "summary": "Get Locations",
        "description": "Get a bounded legacy marker list from Supabase.",
        "operationId": "get_locations_api_v1_gallery_locations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of legacy marker results",
              "default": 500,
              "title": "Limit"
            },
            "description": "Maximum number of legacy marker results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatLocation"
                  },
                  "title": "Response Get Locations Api V1 Gallery Locations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gallery/viewport": {
      "get": {
        "tags": [
          "Gallery"
        ],
        "summary": "Get Locations In Viewport",
        "description": "Get cat locations within a geographic viewport (bounding box).",
        "operationId": "get_locations_in_viewport_api_v1_gallery_viewport_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "north",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "North latitude bound",
              "title": "North"
            },
            "description": "North latitude bound"
          },
          {
            "name": "south",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "South latitude bound",
              "title": "South"
            },
            "description": "South latitude bound"
          },
          {
            "name": "east",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "East longitude bound",
              "title": "East"
            },
            "description": "East longitude bound"
          },
          {
            "name": "west",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "West longitude bound",
              "title": "West"
            },
            "description": "West longitude bound"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GalleryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gallery/search": {
      "get": {
        "tags": [
          "Gallery"
        ],
        "summary": "Search Locations",
        "description": "Search cat locations with optional text query and/or tag filters.",
        "operationId": "search_locations_api_v1_gallery_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text to search in location name and description",
              "title": "Q"
            },
            "description": "Text to search in location name and description"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of tags to filter by",
              "title": "Tags"
            },
            "description": "Comma-separated list of tags to filter by"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page number (alternative to offset)",
              "title": "Page"
            },
            "description": "Page number (alternative to offset)"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SortField"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Sort field",
              "title": "Sort"
            },
            "description": "Sort field"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Sort order",
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gallery/popular-tags": {
      "get": {
        "tags": [
          "Gallery"
        ],
        "summary": "Get Popular Tags",
        "description": "Get the most popular tags used across all cat photos.",
        "operationId": "get_popular_tags_api_v1_gallery_popular_tags_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of top tags to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of top tags to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PopularTagsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gallery/{photo_id}": {
      "get": {
        "tags": [
          "Gallery"
        ],
        "summary": "Get Photo",
        "description": "Get a specific photo by its ID.",
        "operationId": "get_photo_api_v1_gallery__photo_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "The ID of the photo",
              "description": "Must be a valid UUID"
            },
            "description": "Must be a valid UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatLocation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Gallery"
        ],
        "summary": "Delete Photo",
        "description": "Delete a photo. Returns 202 Accepted (deletion runs in background).",
        "operationId": "delete_photo_api_v1_gallery__photo_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "The ID of the photo",
              "description": "Must be a valid UUID"
            },
            "description": "Must be a valid UUID"
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Photo Api V1 Gallery  Photo Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedGalleryResponse": {
        "properties": {
          "images": {
            "items": {
              "$ref": "#/components/schemas/CatLocation"
            },
            "type": "array",
            "title": "Images"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "images",
          "pagination"
        ],
        "title": "PaginatedGalleryResponse",
        "description": "Unified response for offset-based gallery."
      },
...
      "SortField": {
        "type": "string",
        "enum": [
          "uploaded_at",
          "likes_count",
          "comments_count"
        ],
        "title": "SortField",
        "description": "Allowed sort fields for gallery."
      },
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "description": "Sort order."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}